organize ActivityTab code
